pyproject.toml
[project]
name = "quick2-agent"
version = "0.1.0"
description = "Python/FastAPI microservices monorepo for agent orchestration"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "redis>=5.0.0",
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.2.3",
    "sqlalchemy>=2.0.23",
    "apscheduler>=3.10.4",
    "jinja2>=3.1.2",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-instrumentation-fastapi>=0.42b0",
    "httpx>=0.25.0",
    "pyyaml>=6.0.1",
    "minio>=7.2.0",
    "prometheus-client>=0.19.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
    "types-pyyaml>=6.0.12",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["libs*", "services*"]
exclude = ["tests*", "attached_assets*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
